Added TextOverlay and AutoStoryMaker
- TextOverlay takes the working code from the previous commit and abstracts it out of StoryFrameDrawer into its own class. It allows the StoryFrameDrawer code to be much cleaner.
- AutoStoryMaker, at its simplest, takes a story name and generates a video. It has configurable parameters after construction. It mostly takes SampleStory and makes it a little more versatile.
- StoryBitmapManager was split by putting the caching mechanism directly in StoryPage (since it was the only class to use it) and the generalized helper method in the graphics package as BitmapHelper.
- I started adding some code to generate Ken Burns effects in KenBurnsEffectHelper.
- Some of the changes are minor and not worth noting.

One problem I faced (that I think I have had before) was the video encoding process of the MediaCodec with Surface input (PipedVideoSurfaceEncoder) randomly crashing (with IllegalStateException error from dequeueOutputBuffer) after some time. I initially thought this had to do with the differing pixel counts of the source rectangle stretched to the dest rectangle of the surface's canvas because it only happened on the bigger of the two images. However, I figured out that the problem was again just a bitrate problem (I had forgotten to set the bitrate default in AutoStoryMaker).