dramatize playback of drafts
Issue #63
I am still pretty sure that the current functionality of this button is that it stops the audio and restarts it upon hitting play again, but changed the icon to a pause anyway. In the virtual android devices, we can't actually record sound with the recorder so all I have as evidence that it is acting as a stop button is the timing. When I record a 12 sec draft on slide 1 and go to slide 1 in dramatize and hit play and pause after 6 seconds, the second play back is 12 seconds long, which leads me to believe that this is truly acting as a stop button.