Pause StoryPlayer when Clicking Record Button
"In the Learn phase, the playback of the source clip should automatically stop when the microphone is tapped for recording, but this APK allowed simultaneous playing of the original source clip while also recording. This is a bug. I expect when the microphone is tapped that the source playback stops" - Robin